Fix `className` error
Summary:
JavaScript and JSX use `className` to avoid conflicts with the `class`
keyword. This commit fixes an erroneous usage of `class`, which was
causing a React error.

Test Plan:
Patch in #1, then run `yarn start` and open the site in a browser. Note
an error “Invalid DOM property `class`. Did you mean `className`?”
before this commit, and the absence of such an error after it.

Also, run `git grep -F 'class='` and note that there are now no results.

wchargin-branch: fix-classname