Replace hero image `<canvas>` with CSS animation
Summary:
Using CSS animations has a number of benefits over a canvas:

  - The foundation is a vector format, not a raster format, so it
    renders correctly on all screen sizes and zoom settings.
  - Painting and interpolation is automatically handled by the browser.
    It’ll try to hit 60fps, but will throttle down correctly if it can’t
    hit that (rather than lagging).
  - CSS animations support easing, so changing from linear interpolation
    to smooth interpolation is a one-line change (and still has a lot of
    flexibility: see <https://easings.net/>).
  - The other logos use SVG, so now they can all share one codepath.
  - The code is significantly shorter, simpler, and more declarative.
    It’s about a dozen lines of CSS and a dozen more of JavaScript.
  - It doesn’t need JavaScript to run. No `componentDidMount`, either.

I’ve tried to emulate the old canvas logo’s structure and animated
qualities in this new version (though I did fix the ugly linear
interpolation). It’s pretty close, though not identical. Tweak to taste.

Test Plan:
A side-by-side screenshot of the canvas (left) and SVG animation (right)
clearly shows the quality difference; note pixelation on the left:

![Side-by-side still screenshot of the old (left) and new (right)][ss]

[ss]: https://user-images.githubusercontent.com/4317806/63208450-0cd74d80-c089-11e9-9701-fa127d7b6ec4.png

This video is of lower image quality, but shows the improved dynamics;
note on the left side a sharp transition when the first derivative of
the radius changes sign, compared to a smooth curve on the right side:
see [gfycat preview][gfy] or [download high-quality `.webm`][webm].

[gfy]: https://gfycat.com/sizzlinghonorablegoat
[webm]: https://github.com/sourcecred/website/files/3511777/logo_comparison.webm.gz

wchargin-branch: logo-css-animation